Replace dummy translation string
The word "Logo" was the only translated string in the plugin. A dummy
(and never registered) text domain was used.

Remove the translation at all, it does not worth to add translations for
just one word. A filter will do the trick.

Moreover, change default text to "User Image" in place of "Logo".