Various fix in AvatarReplacementService
- undefined var access
- return null when no user meta is stored
- accept as high-res srcset images that are exactly double-sized (and
not only if they are more than that)
- never return the <img> tag wrapped in a url. Reasons are
- `get_avatar()` is more than often used inside anchor tags, and in that
case returnin an `<a>` tag produces an invalid HTML markup (nested <a>
tag are not allowed). Considering it is impossible to predict where the
function is used we should just return the `<img>` tag.
- This plugin want to act as a dropin replacement for standard
`get_avatar()`, so  we should return something fully compatible with
original returnd value, it means just the `<img>` tag